+++ lib/store/my_user.dart
@@ -39,6 +39,8 @@ import '/domain/repository/my_user.dart';
 import '/provider/gql/exceptions.dart';
 import '/provider/gql/graphql.dart';
 import '/provider/hive/my_user.dart';
+import '/util/backoff.dart';
+import '/util/log.dart';
 import '/util/new_type.dart';
 import '/util/platform_utils.dart';
 import '/util/stream_utils.dart';
@@ -96,6 +98,8 @@ class MyUserRepository implements AbstractMyUserRepository {
     required Function() onUserDeleted,
     required Function() onPasswordUpdated,
   }) async {
+    Log.debug('init(onUserDeleted, onPasswordUpdated)', '$runtimeType');
+
     this.onPasswordUpdated = onPasswordUpdated;
     this.onUserDeleted = onUserDeleted;
 
@@ -124,6 +128,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   void dispose() {
+    Log.debug('dispose()', '$runtimeType');
+
     _localSubscription?.cancel();
     _remoteSubscription?.close(immediate: true);
     _keepOnlineSubscription?.cancel();
@@ -131,10 +137,15 @@ class MyUserRepository implements AbstractMyUserRepository {
   }
 
   @override
-  Future<void> clearCache() => _myUserLocal.clear();
+  Future<void> clearCache() async {
+    Log.debug('clearCache()', '$runtimeType');
+    await _myUserLocal.clear();
+  }
 
   @override
   Future<void> updateUserName(UserName? name) async {
+    Log.debug('updateUserName($name)', '$runtimeType');
+
     final UserName? oldName = myUser.value?.name;
 
     myUser.update((u) => u?.name = name);
@@ -149,6 +160,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> updateUserStatus(UserTextStatus? status) async {
+    Log.debug('updateUserStatus($status)', '$runtimeType');
+
     final UserTextStatus? oldStatus = myUser.value?.status;
 
     myUser.update((u) => u?.status = status);
@@ -163,6 +176,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> updateUserLogin(UserLogin login) async {
+    Log.debug('updateUserLogin($login)', '$runtimeType');
+
     final UserLogin? oldLogin = myUser.value?.login;
 
     myUser.update((u) => u?.login = login);
@@ -177,6 +192,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> updateUserPresence(Presence presence) async {
+    Log.debug('updateUserPresence($presence)', '$runtimeType');
+
     final Presence? oldPresence = myUser.value?.presence;
 
     myUser.update((u) => u?.presence = presence);
@@ -194,6 +211,8 @@ class MyUserRepository implements AbstractMyUserRepository {
     UserPassword? oldPassword,
     UserPassword newPassword,
   ) async {
+    Log.debug('updateUserPassword(***, ***)', '$runtimeType');
+
     final bool? hasPassword = myUser.value?.hasPassword;
 
     myUser.update((u) => u?.hasPassword = true);
@@ -210,10 +229,15 @@ class MyUserRepository implements AbstractMyUserRepository {
   }
 
   @override
-  Future<void> deleteMyUser() => _graphQlProvider.deleteMyUser();
+  Future<void> deleteMyUser() async {
+    Log.debug('deleteMyUser()', '$runtimeType');
+    await _graphQlProvider.deleteMyUser();
+  }
 
   @override
   Future<void> deleteUserEmail(UserEmail email) async {
+    Log.debug('deleteUserEmail($email)', '$runtimeType');
+
     if (myUser.value?.emails.unconfirmed == email) {
       final UserEmail? unconfirmed = myUser.value?.emails.unconfirmed;
 
@@ -246,6 +270,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> deleteUserPhone(UserPhone phone) async {
+    Log.debug('deleteUserPhone($phone)', '$runtimeType');
+
     if (myUser.value?.phones.unconfirmed == phone) {
       final UserPhone? unconfirmed = myUser.value?.phones.unconfirmed;
 
@@ -278,6 +304,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> addUserEmail(UserEmail email) async {
+    Log.debug('addUserEmail($email)', '$runtimeType');
+
     final UserEmail? unconfirmed = myUser.value?.emails.unconfirmed;
 
     myUser.update((u) => u?.emails.unconfirmed = email);
@@ -292,6 +320,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> addUserPhone(UserPhone phone) async {
+    Log.debug('addUserPhone($phone)', '$runtimeType');
+
     final UserPhone? unconfirmed = myUser.value?.phones.unconfirmed;
 
     myUser.update((u) => u?.phones.unconfirmed = phone);
@@ -306,6 +336,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> confirmEmailCode(ConfirmationCode code) async {
+    Log.debug('confirmEmailCode($code)', '$runtimeType');
+
     final UserEmail? unconfirmed = myUser.value?.emails.unconfirmed;
 
     await _graphQlProvider.confirmEmailCode(code);
@@ -323,6 +355,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> confirmPhoneCode(ConfirmationCode code) async {
+    Log.debug('confirmPhoneCode($code)', '$runtimeType');
+
     final UserPhone? unconfirmed = myUser.value?.phones.unconfirmed;
 
     await _graphQlProvider.confirmPhoneCode(code);
@@ -339,13 +373,21 @@ class MyUserRepository implements AbstractMyUserRepository {
   }
 
   @override
-  Future<void> resendEmail() => _graphQlProvider.resendEmail();
+  Future<void> resendEmail() async {
+    Log.debug('resendEmail()', '$runtimeType');
+    await _graphQlProvider.resendEmail();
+  }
 
   @override
-  Future<void> resendPhone() => _graphQlProvider.resendPhone();
+  Future<void> resendPhone() async {
+    Log.debug('resendPhone()', '$runtimeType');
+    await _graphQlProvider.resendPhone();
+  }
 
   @override
   Future<void> createChatDirectLink(ChatDirectLinkSlug slug) async {
+    Log.debug('createChatDirectLink($slug)', '$runtimeType');
+
     final ChatDirectLink? link = myUser.value?.chatDirectLink;
 
     myUser.update((u) => u?.chatDirectLink = ChatDirectLink(slug: slug));
@@ -360,6 +402,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> deleteChatDirectLink() async {
+    Log.debug('deleteChatDirectLink()', '$runtimeType');
+
     final ChatDirectLink? link = myUser.value?.chatDirectLink;
 
     myUser.update((u) => u?.chatDirectLink = null);
@@ -377,6 +421,8 @@ class MyUserRepository implements AbstractMyUserRepository {
     NativeFile? file, {
     void Function(int count, int total)? onSendProgress,
   }) async {
+    Log.debug('updateAvatar($file, onSendProgress)', '$runtimeType');
+
     dio.MultipartFile? upload;
 
     if (file != null) {
@@ -429,6 +475,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> toggleMute(MuteDuration? mute) async {
+    Log.debug('toggleMute($mute)', '$runtimeType');
+
     final MuteDuration? muted = myUser.value?.muted;
 
     final Muting? muting = mute == null
@@ -450,6 +498,8 @@ class MyUserRepository implements AbstractMyUserRepository {
     NativeFile? file, {
     void Function(int count, int total)? onSendProgress,
   }) async {
+    Log.debug('updateCallCover($file, onSendProgress)', '$runtimeType');
+
     dio.MultipartFile? upload;
 
     if (file != null) {
@@ -502,6 +552,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> refresh() async {
+    Log.debug('refresh()', '$runtimeType');
+
     final response = await _graphQlProvider.getMyUser();
 
     if (response.myUser != null) {
@@ -511,6 +563,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Initializes [MyUserHiveProvider.boxEvents] subscription.
   Future<void> _initLocalSubscription() async {
+    Log.debug('_initLocalSubscription()', '$runtimeType');
+
     _localSubscription = StreamIterator(_myUserLocal.boxEvents);
     while (await _localSubscription!.moveNext()) {
       BoxEvent event = _localSubscription!.current;
@@ -529,6 +583,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Initializes [_myUserRemoteEvents] subscription.
   Future<void> _initRemoteSubscription() async {
+    Log.debug('_initRemoteSubscription()', '$runtimeType');
+
     _remoteSubscription?.close(immediate: true);
     _remoteSubscription =
         StreamQueue(_myUserRemoteEvents(() => _myUserLocal.myUser?.ver));
@@ -541,6 +597,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Initializes the [GraphQlProvider.keepOnline] subscription.
   void _initKeepOnlineSubscription() {
+    Log.debug('_initKeepOnlineSubscription()', '$runtimeType');
+
     _keepOnlineSubscription?.cancel();
     _keepOnlineSubscription = _graphQlProvider.keepOnline().listen(
       (_) {
@@ -554,6 +612,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Saves the provided [user] in [Hive].
   void _setMyUser(HiveMyUser user, {bool ignoreVersion = false}) {
+    Log.debug('_setMyUser($user, $ignoreVersion)', '$runtimeType');
+
     if (user.ver > _myUserLocal.myUser?.ver || ignoreVersion) {
       user.value.blocklistCount ??= _myUserLocal.myUser?.value.blocklistCount;
       _myUserLocal.set(user);
@@ -565,9 +625,18 @@ class MyUserRepository implements AbstractMyUserRepository {
     var userEntity = _myUserLocal.myUser;
 
     if (userEntity == null || versioned.ver <= userEntity.ver) {
+      Log.debug(
+        '_myUserRemoteEvent(): ignored ${versioned.events.map((e) => e.kind)}',
+        '$runtimeType',
+      );
       return;
     }
     userEntity.ver = versioned.ver;
+
+    Log.debug(
+      '_myUserRemoteEvent(): ${versioned.events.map((e) => e.kind)}',
+      '$runtimeType',
+    );
 
     for (var event in versioned.events) {
       // Updates a [User] associated with this [MyUserEvent.userId].
@@ -767,32 +836,51 @@ class MyUserRepository implements AbstractMyUserRepository {
   /// Subscribes to remote [MyUserEvent]s of the authenticated [MyUser].
   Stream<MyUserEventsVersioned> _myUserRemoteEvents(
     MyUserVersion? Function() ver,
-  ) =>
-      _graphQlProvider.myUserEvents(ver).asyncExpand((event) async* {
-        var events =
-            MyUserEvents$Subscription.fromJson(event.data!).myUserEvents;
+  ) {
+    Log.debug('_myUserRemoteEvents(ver)', '$runtimeType');
 
-        if (events.$$typename == 'SubscriptionInitialized') {
-          events
-              as MyUserEvents$Subscription$MyUserEvents$SubscriptionInitialized;
-          // No-op.
-        } else if (events.$$typename == 'MyUser') {
-          _setMyUser((events as MyUserMixin).toHive(
-              blocklistCount:
-                  (events as MyUserEvents$Subscription$MyUserEvents$MyUser)
-                      .blocklist
-                      .totalCount));
-        } else if (events.$$typename == 'MyUserEventsVersioned') {
-          var mixin = events as MyUserEventsVersionedMixin;
-          yield MyUserEventsVersioned(
-            mixin.events.map((e) => _myUserEvent(e)).toList(),
-            mixin.ver,
-          );
-        }
-      });
+    return _graphQlProvider.myUserEvents(ver).asyncExpand((event) async* {
+      Log.trace('_myUserRemoteEvents(ver): ${event.data}', '$runtimeType');
+
+      var events = MyUserEvents$Subscription.fromJson(event.data!).myUserEvents;
+
+      if (events.$$typename == 'SubscriptionInitialized') {
+        Log.debug(
+          '_myUserRemoteEvents(ver): SubscriptionInitialized',
+          '$runtimeType',
+        );
+
+        events
+            as MyUserEvents$Subscription$MyUserEvents$SubscriptionInitialized;
+        // No-op.
+      } else if (events.$$typename == 'MyUser') {
+        Log.debug(
+          '_myUserRemoteEvents(ver): MyUser',
+          '$runtimeType',
+        );
+
+        _setMyUser(
+          (events as MyUserMixin).toHive(
+            blocklistCount:
+                (events as MyUserEvents$Subscription$MyUserEvents$MyUser)
+                    .blocklist
+                    .totalCount,
+          ),
+        );
+      } else if (events.$$typename == 'MyUserEventsVersioned') {
+        var mixin = events as MyUserEventsVersionedMixin;
+        yield MyUserEventsVersioned(
+          mixin.events.map((e) => _myUserEvent(e)).toList(),
+          mixin.ver,
+        );
+      }
+    });
+  }
 
   /// Constructs a [MyUserEvent] from the [MyUserEventsVersionedMixin$Events].
   MyUserEvent _myUserEvent(MyUserEventsVersionedMixin$Events e) {
+    Log.trace('_myUserEvent($e)', '$runtimeType');
+
     if (e.$$typename == 'EventUserNameUpdated') {
       var node = e as MyUserEventsVersionedMixin$Events$EventUserNameUpdated;
       return EventUserNameUpdated(node.userId, node.name);