+++ lib/ui/page/home/page/chat/controller.dart
@@ -1332,7 +1332,7 @@ class ChatController extends GetxController {
   /// Loads previous page of the [RxChat.messages], if [_atTop].
   Future<void> _loadPreviousPage() async {
     if (hasPrevious.isTrue && chat!.previousLoading.isFalse && _atTop) {
-      Log.print('Fetch previous page', 'ChatController');
+      Log.info('Fetch previous page', 'ChatController');
 
       if (_topLoader == null) {
         _topLoader = LoaderElement.top();